made hardware.hpp, renamed stuff
