+++ chip8.hpp
@@ -1,46 +1,15 @@
+#ifndef CHIP8HPP
+#define CHIP8HPP
 #include <random>
-#include <thread>
-#include <mutex>
-#include <thread>
 #include <chrono>
 #include <cstdint>
+#include <exception>
+#include <iostream>
+#include <random>
+#include <string>
+#include "hardware.hpp"
 
 constexpr std::chrono::nanoseconds CYCLE_TIME(1666667);
-
-// Implements Chip 8 style memory
-class MemoryFile {
-public:
-    MemoryFile();
-
-    uint8_t& operator[](uint16_t addr);
-private:
-    uint8_t mem[4096];
-};
-
-// Implements a Chip 8 style display
-class Display {
-public:
-    Display();
-    
-    void clear();
-    uint8_t drawsprite(int x, int y, int len, const uint8_t *sprite);
-    void updatescreen();
-private:
-    uint64_t display[32];
-};
-
-class RegisterFile {
-public:
-    RegisterFile();
-
-    uint8_t& operator[](int regIndex) { return registers[regIndex]; }
-    uint16_t& I() { return memReg; }
-    void saveRange(int lastReg, uint16_t addr, MemoryFile& mem);
-    void loadRange(int lastReg, uint16_t addr, MemoryFile& mem);
-private:
-    uint8_t registers[16];
-    uint16_t memReg;
-};
 
 // The Chip 8 interpreter
 class Chip8 {
@@ -79,3 +48,5 @@ private:
     std::mt19937 re;
     std::uniform_int_distribution<std::mt19937::result_type> dist; // distribution in range [1, 6]
 };
+
+#endif
